+++ src/ts/Vertex.ts
@@ -25,7 +25,8 @@
  * @modified 2021-12-01 Changed the type of param of `scale` to XYCoords.
  * @modified 2021-12-01 Added function `scaleXY` for non uniform scaling.
  * @modified 2021-12-17 Added the functions `lerp` and `lerpAbs` for linear interpolations.
- * @version  2.6.0
+ * @modified 2022-01-31 Added `Vertex.utils.arrayToJSON`.
+ * @version  2.6.1
  *
  * @file Vertex
  * @public
@@ -647,6 +648,22 @@ export class Vertex implements XYCoords, SVGSerializable {
       );
 
       return vertices;
+    },
+
+    /**
+     * Convert the given vertices (array) to a JSON string.
+     *
+     * @param {number?} precision - (optional) The numeric precision to be used (number of precision digits).
+     * @returns {string}
+     */
+    arrayToJSON(vertices: Array<XYCoords>, precision?: number) {
+      return JSON.stringify(
+        vertices.map(function (vert) {
+          return typeof precision === undefined
+            ? { x: vert.x, y: vert.y }
+            : { x: vert.x.toFixed(precision), y: vert.y.toFixed(precision) };
+        })
+      );
     }
   };
 }